Add files via upload
Tennis Pong Game made for all age groups--

How to Play !!!
The left side paddle is for the user and the right side paddle is for the computer. The game automatically starts when you load the page. Move mouse up/down to move left user's paddle vertically within the game area. If computer misses the ball user gets the point and vice-versa. Score 10 points first to win the game.